// src/interfaces/messagesType.ts
export enum MessagesType {
    GET_READER_CAPABILITIES = 1,
    GET_READER_CAPABILITIES_RESPONSE = 11,
    ADD_ROSPEC = 20,
    ADD_ROSPEC_RESPONSE = 30,
    DELETE_ROSPEC = 21,
    DELETE_ROSPEC_RESPONSE = 31,
    START_ROSPEC = 22,
    START_ROSPEC_RESPONSE = 32,
    STOP_ROSPEC = 23,
    STOP_ROSPEC_RESPONSE = 33,
    ENABLE_ROSPEC = 24,
    ENABLE_ROSPEC_RESPONSE = 34,
    DISABLE_ROSPEC = 25,
    DISABLE_ROSPEC_RESPONSE = 35,
    GET_ROSPECS = 26,
    GET_ROSPECS_RESPONSE = 36,
    ADD_ACCESSSPEC = 40,
    ADD_ACCESSSPEC_RESPONSE = 50,
    DELETE_ACCESSSPEC = 41,
    DELETE_ACCESSSPEC_RESPONSE = 51,
    ENABLE_ACCESSSPEC = 42,
    ENABLE_ACCESSSPEC_RESPONSE = 52,
    DISABLE_ACCESSSPEC = 43,
    DISABLE_ACCESSSPEC_RESPONSE = 53,
    GET_ACCESSSPECS = 44,
    GET_ACCESSSPECS_RESPONSE = 54,
    CLIENT_REQUEST_OP = 45,
    CLIENT_REQUEST_OP_RESPONSE = 55,
    GET_REPORT = 60,
    RO_ACCESS_REPORT = 61,
    KEEPALIVE = 62,
    KEEPALIVE_ACK = 72,
    READER_EVENT_NOTIFICATION = 63,
    ENABLE_EVENTS_AND_REPORTS = 64,
    ERROR_MESSAGE = 100,
    GET_READER_CONFIG = 2,
    GET_READER_CONFIG_RESPONSE = 12,
    SET_READER_CONFIG = 3,
    SET_READER_CONFIG_RESPONSE = 13,
    CLOSE_CONNECTION = 14,
    CLOSE_CONNECTION_RESPONSE = 4,
    CUSTOM_MESSAGE = 1023
}